+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
  * [BC BREAK] `CachedPlatform` has been renamed `CachePlatform` and moved as a bridge, please require `symfony/ai-cache-platform` and use `Symfony\AI\Platform\Bridge\Cache\CachePlatform`
  * [BC BREAK] `Metadata::merge()` method signature has changed to accept `Metadata` instead of array
  * [BC BREAK] Behavior of `Metadata::add()` has changed to merge existing keys instead of overwriting them
+ * [BC BREAK] Move `Symfony\AI\Platform\Serializer\StructuredOutputSerializer` to `Symfony\AI\Platform\StructuredOutput\Serializer`
 
 0.2
 ---